# perf(compile): pass module source data from binary directly to v8