--- a/cli/standalone/file_system.rs
+++ b/cli/standalone/file_system.rs
@@ -22,8 +22,8 @@ use super::virtual_fs::FileBackedVfs;
 pub struct DenoCompileFileSystem(Arc<FileBackedVfs>);
 
 impl DenoCompileFileSystem {
-  pub fn new(vfs: FileBackedVfs) -> Self {
-    Self(Arc::new(vfs))
+  pub fn new(vfs: Arc<FileBackedVfs>) -> Self {
+    Self(vfs)
   }
 
   fn error_if_in_vfs(&self, path: &Path) -> FsResult<()> {