--- a/cli/standalone/mod.rs
+++ b/cli/standalone/mod.rs
@@ -315,36 +315,19 @@ impl ModuleLoader for EmbeddedModuleLoader {
     }
 
     match self.shared.modules.read(original_specifier) {
-      Ok(Some(module)) => deno_core::ModuleLoadResponse::Sync(Ok(
-        deno_core::ModuleSource::new_with_redirect(
-          match module.media_type {
-            MediaType::JavaScript
-            | MediaType::Jsx
-            | MediaType::Mjs
-            | MediaType::Cjs
-            | MediaType::TypeScript
-            | MediaType::Mts
-            | MediaType::Cts
-            | MediaType::Dts
-            | MediaType::Dmts
-            | MediaType::Dcts
-            | MediaType::Tsx => ModuleType::JavaScript,
-            MediaType::Json => ModuleType::Json,
-            MediaType::Wasm => ModuleType::Wasm,
-            // just assume javascript if we made it here
-            MediaType::TsBuildInfo
-            | MediaType::SourceMap
-            | MediaType::Unknown => ModuleType::JavaScript,
-          },
-          ModuleSourceCode::Bytes(match module.data {
-            Cow::Borrowed(d) => d.into(),
-            Cow::Owned(d) => d.into_boxed_slice().into(),
-          }),
-          original_specifier,
-          module.specifier,
-          None,
-        ),
-      )),
+      Ok(Some(module)) => {
+        let (module_specifier, module_type, module_source) =
+          module.into_for_v8();
+        deno_core::ModuleLoadResponse::Sync(Ok(
+          deno_core::ModuleSource::new_with_redirect(
+            module_type,
+            module_source,
+            original_specifier,
+            module_specifier,
+            None,
+          ),
+        ))
+      }
       Ok(None) => deno_core::ModuleLoadResponse::Sync(Err(type_error(
         format!("{MODULE_NOT_FOUND}: {}", original_specifier),
       ))),