--- a/tests/specs/compile/byonm_main_sub_dir/deno.json
+++ b/tests/specs/compile/byonm_main_sub_dir/deno.json
@@ -1,3 +1,3 @@
 {
-  "unstable": ["byonm"]
+  "nodeModulesDir": "manual"
 }