… Cargo.lock
… cli/Cargo.toml
… cli/standalone/virtual_fs.rs
… cli/tools/compile.rs
… cli/util/text_encoding.rs
… tests/specs/compile/detect_cjs/output.out