This changes `denort` to pass a static reference of the moude source bytes found in the binary to v8 instead of copying it.

## Memory Usage (Example 1)

```ts
async function test() {
  await import("./typescript.js");
  await import("./typescript2.js");
}

setInterval(() => {
  console.log(Deno.memoryUsage().rss);
}, 1000);
```

```shellsession
% ./scratch_main
80805888
80920576
81231872
^C
% ./scratch_new
36044800
36175872
36487168
36487168
36536320
```

## Memory Usage (Example 2)

```ts
import "./typescript.js";
import "./typescript2.js";

setInterval(() => {
  console.log(Deno.memoryUsage().rss);
}, 1000);
```

```shellsession
% ./scratch_main                   
161644544
161742848
161742848
161775616
^C
% ./scratch_new 
97746944
97828864
97845248
97878016
^C
```

## Performance

```
Benchmarking not_imported_dynamic_import.ts
Benchmark 1: ./scratch_new
  Time (mean ± σ):      15.3 ms ±   0.9 ms    [User: 10.6 ms, System: 3.4 ms]
  Range (min … max):    14.1 ms …  20.6 ms    179 runs
 
Benchmark 2: ./scratch_main
  Time (mean ± σ):      19.8 ms ±   0.6 ms    [User: 12.8 ms, System: 5.4 ms]
  Range (min … max):    18.6 ms …  23.0 ms    138 runs
 
Summary
  ./scratch_new ran
    1.30 ± 0.09 times faster than ./scratch_main
Benchmarking typescript_imports.ts
Benchmark 1: ./scratch_new
  Time (mean ± σ):     169.6 ms ±   1.2 ms    [User: 161.4 ms, System: 11.6 ms]
  Range (min … max):   167.9 ms … 171.5 ms    17 runs
 
Benchmark 2: ./scratch_main
  Time (mean ± σ):     176.1 ms ±   2.0 ms    [User: 163.7 ms, System: 15.6 ms]
  Range (min … max):   173.9 ms … 179.9 ms    16 runs
 
Summary
  ./scratch_new ran
    1.04 ± 0.01 times faster than ./scratch_main
Benchmarking many_jsr.ts
Benchmark 1: ./scratch_new
  Time (mean ± σ):      61.6 ms ±   1.4 ms    [User: 56.0 ms, System: 7.4 ms]
  Range (min … max):    59.9 ms …  67.9 ms    47 runs
 
Benchmark 2: ./scratch_main
  Time (mean ± σ):      49.9 ms ±   0.6 ms    [User: 44.8 ms, System: 7.4 ms]
  Range (min … max):    48.6 ms …  51.4 ms    58 runs
 
Summary
  ./scratch_main ran
    1.23 ± 0.03 times faster than ./scratch_new
```

<details>
<summary>Files</summary>

```ts
// not_imported_dynamic_import.ts
console.log("Hello");

async function test() {
  await import("./typescript.js");
  await import("./typescript2.js");
}

// typescript_imports.ts
import "./typescript.js";
import "./typescript2.js";

console.log("hello");

// many_jsr.ts
import "jsr:@std/assert";
import "jsr:@std/async";
import "jsr:@std/bytes";
import "jsr:@std/cache";
import "jsr:@std/collections";
import "jsr:@std/csv";
import "jsr:@std/data-structures";
import "jsr:@std/datetime";
import "jsr:@std/dotenv";
import "jsr:@std/encoding";
import "jsr:@std/expect";
import "jsr:@std/fmt/bytes";
import "jsr:@std/fmt/colors";
import "jsr:@std/fmt/duration";
import "jsr:@std/fmt/printf";
import "jsr:@std/front-matter";
import "jsr:@std/fs";
import "jsr:@std/html";
import "jsr:@std/http";
import "jsr:@std/ini";
import "jsr:@std/io";
import "jsr:@std/jsonc";
import "jsr:@std/log";
import "jsr:@std/media-types";
import "jsr:@std/msgpack";
import "jsr:@std/net";
import "jsr:@std/path";
import "jsr:@std/random";
import "jsr:@std/regexp";
import "jsr:@std/semver";
import "jsr:@std/streams";
import "jsr:@std/tar";
import "jsr:@std/testing/bdd";
import "jsr:@std/testing/mock";
import "jsr:@std/testing/snapshot";
import "jsr:@std/testing/time";
import "jsr:@std/testing/types";
import "jsr:@std/text";
import "jsr:@std/toml";
import "jsr:@std/ulid";
import "jsr:@std/uuid";
import "jsr:@std/yaml";
import "jsr:@david/dax";
console.log("Hello");
```

</details>

It's faster in many scenarios, but seems slower for the test running many JSR modules. I looked at a flamegraph and it seems to be due to https://github.com/denoland/deno_core/issues/942