@KnorpelSenf 2024-10-27
This introduced a bug that I reported in #26583.